+++ packages/components/inputs/radio-input/src/index.js
@@ -2,3 +2,5 @@ import Group from './radio-group';
 import Option from './radio-option';
 
 export default { Group, Option };
+
+export { default as version } from './version';
