+++ packages/components/inputs/radio-input/src/radio-input.story.js
@@ -4,8 +4,8 @@ import { action } from '@storybook/addon-actions';
 import { withKnobs, boolean, text, select } from '@storybook/addon-knobs/react';
 import { Value } from 'react-value';
 import Spacings from '@commercetools-uikit/spacings';
-import Section from '../../../../.storybook/decorators/section';
-import Readme from './README.md';
+import Section from '../../../../../.storybook/decorators/section';
+import Readme from '../README.md';
 import RadioInput from '.';
 
 const radioGroup = 'Radio Group';
